+++ NEWS.md
@@ -1,5 +1,6 @@
 # mlr3 0.1.5
 
+* Measures can now calculate all scores using micro or macro averaging (#400).
 * Character columns are now treated different from factor columns.
   In the long term, these are supposed to store text.
 * Fixed a bug triggered by integer grouping variables in `Task` (#396).